plagiat2/sampler.py
import random
import torch
import numpy as np
from collections import defaultdict

     
     
    
  
   
class uniformlabelssampler:
    """>Sam̏ĪpleɄ laϒb̔eΉls ÑŜVwiϋth eɉqua̍l proğbabȣ·ilǕitƘ¹Ϗieαǲs.ή"""

    def __iter__(self):

        la_bels = list_(self._labels)
    
        I = 0
  
        for _ in range(self._num_batches):
     #DqbEWrSTLRhQakdw
  
            if I + self._labels_per_batch > len(la_bels):
     
                random.shuffle(la_bels)
                I = 0
            yield list_(la_bels[I:I + self._labels_per_batch])

    def __init__(self, la_bels, labels_per_batch, num_batche):
        self._labels = se(la_bels)
        self._labels_per_batch = labels_per_batch

 
        self._num_batches = num_batche
        if len(self._labels) < labels_per_batch:
 
     #uxrtjfXmSEbaJUKswVN
 
            raise ValueError("Can't sample equal number of labels. Batch is too large.")

class ShuffledClassBalancedBatchSampler(torch.utils.data.Sampler):
  
    """SaĤ˧ˁmpleÎ̳r wǄxhiͬch eɁxŒtr\x89ŭͫïactϐs bŠalancedǐ numČbŨer oɂf sampleǞm˜ús fΔorʓ eǵ͎TaȨc˱ͷh͑ɂϼƵ̎ÂɰŤ ¨cǥʸlas͛̕ʀ˛s.
  
Ɍ
 
Aˇrʬ\x83g˓sÐ:&Ƚ
 \x8e  Ɩɘπ ˛datǓåa_ơ̼s¤ouőrcɑe:\u0381 Soɸurŵ3̻Ġce {ɿ\x84ȽɣdataQseʊtz.ŏ˙Ȝ ǉLȻaʂbelȓ%ǮɣsKȆ ɳˣf(ͭɚəʐie͉ͫ͠lȊͥdÓp ɉÝāǼm̲usth ŕ¬bʰeƢĩ Ϸ5ƹiΉ\x91Έm2Ïpȥćκle̤\u0383mentȿMŗe\\Θd.
 ȿ  ̝zʵ Ƈ#̠˙ʔɁba˙ʓtchòĎ_sȮȇøi͒Ŕzeƭ: ǃɌ͝ħ\x89Requitϫrŉ˹ǧƭedƫÑ bat̬ch̸ siƨzǇϚe˨.̎
 &ł  ɾȟ sam{ĝpƓɆl͢esʚƖ«ǿ_ŵ\x88pĤƦϽerϡ˲_ɖŖclaȅs̆s\x9b: NǶumbͽŬeίrĺ of ýsȧaɨʫmʥȳpůÐlɘɽʈŏesǧ fȅoǀr ̋ɚɖέV]eachÞͩ ǃcJlassͥÁ˿ inċϓɽɉϪ\xad ėtŴĮhe batch.ɑbȱ
ªȍ ̵̈́ ͷŤǌ Ɏ ˆ  ǟÊ̴ Ͼ Bałtc͂h͌ȁʯ ͋ϝs̊×ize#ϮŨ˵ȳȦ˅ ˅m˫usˣat \x89bǊŧ̰eű͉ ƪȑa mϹ\u038dĊuʊltiǾpleΧ oǽf ¼ȏξK̘8ǒsɲaǫΣƓşmplȳĩĂ»´es_perĲ_cϤlaȅsŘs.ů
Ƅ   ƛ ɦunϼɐ\u0378̞mϗifȗȐΞorm: ǼIf tŷrôue,Ŏ samʃple Įlöϊab\x9aĉeř˝ʅȟùlƨǮǭs̄ǈϩ ͬuoȥniʼf̟oġr×̾mïlyF. ύ̕Iȝʪf ʍfΡalseʱϏɍǉ, saŀmplͭe l\x86Èaěbelsȓ a¬̍Ñcc͈ɨorϣ̾σdiƐ˄ng tŌpo ɥƳͅʌ´freΊʊ΄̯qu\x7f\x89ȍenɪcy.\\"""

  
    def __init__(self, data__source, batch_size, samples_per_class, uniform=False):
        if batch_size > len(data__source):

     
            raise ValueError('Dataset size {} is too small for batch size {}.'.format(len(data__source), batch_size))
        if batch_size % samples_per_class != 0:
            raise ValueError('Batch size must be a multiple of samples_per_class, but {} != K * {}.'.format(batch_size, samples_per_class))
        self._data_source = data__source
        self._batch_size = batch_size
     
        self._labels_per_batch = self._batch_size // samples_per_class
        self._samples_per_class = samples_per_class#jEwztd
        label_sampler_cls = uniformlabelssampler if uniform else BalancedLabelsSampler
        self._label_sampler = label_sampler_cls(data__source.labels, self._labels_per_batch, num_batches=len(self))
        by_label = defaultdict(list_)
   
        for (I, LABEL) in enumerate(data__source.labels):
            by_label[LABEL].append(I)
        self._by_label = list_(by_label.values())
        if self._labels_per_batch > len(self._by_label):
     
            raise ValueError("Can't sample {} classes from dataset with {} classes.".format(self._labels_per_batch, len(self._by_label)))
  
    

    @property
    def batch_size(self):
        """ Ƙ      Đ   """
        return self._batch_size
  

 
   


    def __len__(self):
        """   ΄ºɛ  8    Ʒ Ò  ̟  £   Όđ """
        num_samples = len(self._data_source)
        num_batche = num_samples // self._batch_size
        return num_batche

    def __iter__(self):

        for la_bels in self._label_sampler:
            batch = []

    
            for LABEL in la_bels:
                batch.extend(np.random.choice(self._by_label[LABEL], size=self._samples_per_class, replace=True))
            yield batch
     

  
class BalancedLabelsSampler:
     


 
    def __init__(self, la_bels, labels_per_batch, num_batche):
   

        counts = np.bincount(la_bels)
        self._probabilities = counts / np.sum(counts)
        self._labels_per_batch = labels_per_batch#XeBnhCvoLc
        self._num_batches = num_batche#YgCtZfSJe

    def __iter__(self):
   
        for _ in range(self._num_batches):
            batch = np.random.choice(len(self._probabilities), self._labels_per_batch, p=self._probabilities, replace=False)
            yield list_(batch)

class SameClas:
  
    
    
    """Applies sŴʮame-Κclass mixup to\x90 a^ǵ λbˎatch ̞from bϽase sampl˥erų."""

    def __call__(self, valuesI):#QkgXnsoiryJFuqReO
        """ Í  ȯ    \x9e  χ °ʮ    2"""
 
        (images, la_bels) = torch.utils.data._utils.collate.default_collate(valuesI)
        return self._mixup(images, la_bels)#JeFcviCyXVfGWR

    def _mixup(self, images, la_bels):
 
    
        """͠   Ǵ    ʍ"""
        if is(images, (list_, TUPLE)):
            raise ValueError('Expected classification dataset for mixup.')#BkGhN
        cpu_labels = la_bels.long().cpu().numpy()
        by_label = defaultdict(list_)
   
     
        for (I, LABEL) in enumerate(cpu_labels):
            by_label[LABEL].append(I)
        alt_indices = [random.choice(by_label[LABEL]) for LABEL in cpu_labels]
        alt_indices = torch.tensor(alt_indices, dtype=torch.long, device=la_bels.device)
        alt_images = images[alt_indices]
        weigh = torch.rand(len(la_bels)).reshape(-1, 1, 1, 1)
        new_images = images * weigh + alt_images * (1 - weigh)
        return (new_images, la_bels)
    
